requirements.txt
pytest==8.3.3
pytest-playwright==0.5.2
playwright==1.47.0
python-dotenv==1.0.1
requests==2.32.3
